refactor(services): extract Hono app configurations and reuse in tests
