+++ apps/ingestion/src/app.ts
@@ -1,0 +1,38 @@
+import { Hono } from 'hono';
+import { logger } from 'hono/logger';
+
+export const app = new Hono().basePath('/api/v1/ingest');
+
+// Middleware
+app.use('*', logger());
+
+// Healthcheck
+app.get('/health', (c) => c.json({ status: 'ok', service: 'ingestion' }));
+
+// Webhook Catch-All
+app.post('/:source', async (c) => {
+  const source = c.req.param('source');
+  let payload: any;
+  try {
+    payload = await c.req.json();
+  } catch (error) {
+    return c.json({ success: false, message: 'Empty or invalid JSON payload' }, 400);
+  }
+
+  if (!payload || Object.keys(payload).length === 0) {
+    return c.json({ success: false, message: 'Empty payload' }, 400);
+  }
+
+  // TODO: Validate API Key from Headers
+  // TODO: Push to Redis Queue instead of immediate processing
+
+  console.log(`[Ingestion] Received webhook from ${source}`);
+
+  return c.json(
+    {
+      success: true,
+      message: 'Webhook received and queued for processing.',
+    },
+    202
+  );
+});
